+++ Models/Requirement.cs
@@ -9,9 +9,9 @@ namespace SystemChecker.Models
         public List<string> Os { get; set; } = new List<string>();
         public List<string> Cpu { get; set; } = new List<string>();
         public List<string> Gpu { get; set; } = new List<string>();
-        public long Ram { get; set; }
+        public double Ram { get; set; }
         public string RamUsage { get; set; }
-        public long Storage { get; set; }
+        public double Storage { get; set; }
         public string StorageUsage { get; set; }
 
         [JsonIgnore]
